Добавлен Insert запрос для таблиц.

+++ app/src/main/java/ru/mertsalovda/roomdemoapp/MusicProvider.java
@@ -24,12 +24,14 @@ public class MusicProvider extends ContentProvider {
     private static final String TABLE_SONG = "song";
     private static final String TABLE_ALBUMSONG = "albumsong";
 
+    public static final String KEY_DATA_ID = "KEY_DATA_ID";
+    public static final String KEY_DATA_2 = "KEY_DATA_2";
+    public static final String KEY_DATA_3 = "KEY_DATA_3";
+
     private static final UriMatcher URI_MATCHER = new UriMatcher(UriMatcher.NO_MATCH);
 
     private static final int ALBUM_TABLE_CODE = 100;
     private static final int ALBUM_ROW_CODE = 101;
-    private static final int ALBUM_TABLE_CODE_ID = 110;
-    private static final int ALBUM_ROW_CODE_ID = 111;
 
     private static final int SONG_TABLE_CODE = 200;
     private static final int SONG_ROW_CODE = 201;
@@ -126,8 +128,54 @@ public class MusicProvider extends ContentProvider {
 
     @Override
     public Uri insert(Uri uri, ContentValues values) {
+        int code = URI_MATCHER.match(uri);
+        if (code != ALBUM_ROW_CODE && code != ALBUM_TABLE_CODE
+                && code != SONG_ROW_CODE && code != SONG_TABLE_CODE
+                && code != ALBUMSONG_ROW_CODE && code != ALBUMSONG_TABLE_CODE) return null;
+        int _ID;
+        String name;
+        switch (code) {
+            case ALBUM_TABLE_CODE:
+                _ID = values.getAsInteger(KEY_DATA_ID);
+                if (mMusicDao.getAlbumById(_ID) == null) {
+                    name = values.getAsString(KEY_DATA_2);
+                    String reales = values.getAsString(KEY_DATA_3);
+                    mMusicDao.insertAlbum(new Album(_ID, name, reales));
+                } else {
+                    uri = Uri.parse("");
+                }
+                break;
+            case SONG_TABLE_CODE:
+                _ID = values.getAsInteger(KEY_DATA_ID);
+                if (mMusicDao.getSongById(_ID) == null) {
+                    name = values.getAsString(KEY_DATA_2);
+                    String duration = values.getAsString(KEY_DATA_3);
+                    mMusicDao.insertSong(new Song(_ID, name, duration));
+                } else {
+                    uri = Uri.parse("");
+                }
+                break;
+            case ALBUMSONG_TABLE_CODE:
+                _ID = values.getAsInteger(KEY_DATA_ID);
+                int album_id = values.getAsInteger(KEY_DATA_2);
+                int song_id = values.getAsInteger(KEY_DATA_3);
+                if (mMusicDao.getAlbumSongById(_ID) == null
+                        && AlbumAndSongIsNotNull(album_id, song_id)) {
+                    mMusicDao.insertAlbumSong(new AlbumSong(_ID, album_id, song_id));
+                } else {
+                    uri = Uri.parse("");
+                }
+                break;
+        }
+        return uri;
+    }
 
-        throw new UnsupportedOperationException("Not yet implemented");
+    private boolean AlbumAndSongIsNotNull(int album_id, int song_id) {
+        boolean result = false;
+        if (mMusicDao.getAlbumById(album_id) != null
+                && mMusicDao.getSongById(song_id) != null)
+            result = true;
+        return result;
     }
 
     @Override
@@ -148,29 +196,41 @@ public class MusicProvider extends ContentProvider {
         List<AlbumSong> albumSongs;
         switch (code) {
             case ALBUM_ROW_CODE:
-                Album album = mMusicDao.getAlbumsById(id);
-                albumSongs = mMusicDao.getAlbumsFromAlbumSong(id);
-                for (AlbumSong as: albumSongs){
-                    mMusicDao.deleteAlbumSong(as);
+                Album album = mMusicDao.getAlbumById(id);
+                if (album != null) {
+                    albumSongs = mMusicDao.getAlbumsFromAlbumSong(id);
+                    for (AlbumSong as : albumSongs) {
+                        mMusicDao.deleteAlbumSong(as);
+                        deleted++;
+                    }
+                    mMusicDao.deleteAlbum(album);
                     deleted++;
+                } else {
+                    deleted = -1;
                 }
-                mMusicDao.deleteAlbum(album);
-                deleted++;
                 break;
             case SONG_ROW_CODE:
                 Song song = mMusicDao.getSongById(id);
-                albumSongs = mMusicDao.getSongsFromAlbumSong(id);
-                for (AlbumSong as: albumSongs){
-                    mMusicDao.deleteAlbumSong(as);
+                if (song != null) {
+                    albumSongs = mMusicDao.getSongsFromAlbumSong(id);
+                    for (AlbumSong as : albumSongs) {
+                        mMusicDao.deleteAlbumSong(as);
+                        deleted++;
+                    }
+                    mMusicDao.deleteSong(song);
                     deleted++;
+                } else {
+                    deleted = -1;
                 }
-                mMusicDao.deleteSong(song);
-                deleted++;
                 break;
             case ALBUMSONG_ROW_CODE:
                 AlbumSong albumSong = mMusicDao.getAlbumSongById(id);
-                mMusicDao.deleteAlbumSong(albumSong);
-                deleted++;
+                if (albumSong != null) {
+                    mMusicDao.deleteAlbumSong(albumSong);
+                    deleted++;
+                } else {
+                    deleted = -1;
+                }
                 break;
         }
         return deleted;
